Lesson 012 - Numbers; float

+++ numbers.py
@@ -1,3 +1,3 @@
 age = input("Enter your age: ")
-new_age = int(age) + 50
+new_age = float(age) + 50
 print(new_age)
